% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_references.R
\name{build_Athanasiadou_reference}
\alias{build_Athanasiadou_reference}
\title{Generate differential expression reference for Athanasiadou et al. (2021) 
bulk RNA-seq data}
\usage{
build_Athanasiadou_reference()
}
\value{
named list of observed counts in conditions 1 and 2
}
\description{
Generate differential expression reference for Athanasiadou et al. (2021) 
bulk RNA-seq data
}
