% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{build_theta}
\alias{build_theta}
\title{Simulate differential abundance parameters}
\usage{
build_theta(beta.0, proportion_da, spike_in)
}
\arguments{
\item{beta.0}{vector of log baseline (mean) expression for all genes}

\item{proportion_da}{proportion of differentially abundant genes}

\item{spike_in}{flag indicating whether or not to simulate a spike-in (control) with very low dispersion}
}
\value{
complete vector of dispersion and log coefficient parameters
}
\description{
Simulate differential abundance parameters
}
