% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{calc_DA_discrepancy}
\alias{calc_DA_discrepancy}
\title{Calculates the discrepancy between differential abundance calls made on
absolute and relative abundances}
\usage{
calc_DA_discrepancy(calls, oracle_calls)
}
\arguments{
\item{calls}{calls (p-value vector) made on relative abundances (the observed 
data)}

\item{oracle_calls}{calls (p-value vector) made on the absolute abundances}
}
\value{
named list of true positive rate, false positive rate with and
without multiple test correction
}
\description{
Calculates the discrepancy between differential abundance calls made on
absolute and relative abundances
}
