% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{calc_DE_discrepancy}
\alias{calc_DE_discrepancy}
\title{Calculates the discrepancy between differential abundance calls made on
absolute and relative abundances}
\usage{
calc_DE_discrepancy(
  ref_data,
  data,
  groups,
  method = "NBGLM",
  oracle_calls = NULL
)
}
\arguments{
\item{ref_data}{absolute abundance data set (samples x features)}

\item{data}{relative abundance data set (samples x features)}

\item{groups}{group (cohort) labels}

\item{method}{differential abundance calling method (e.g. "DESeq2")}

\item{oracle_calls}{optional baseline (true) differential abundance calls}
}
\value{
named list of true positive rate, false positive rate, and baseline
(true) differential expression calls
}
\description{
Calculates the discrepancy between differential abundance calls made on
absolute and relative abundances
}
