% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBID_optimization.R
\name{calc_LRT.NBID}
\alias{calc_LRT.NBID}
\title{Calculate p-value from likelihood ratio test over NBID full and null models}
\usage{
calc_LRT.NBID(theta_H0, theta_H1, other_params)
}
\arguments{
\item{theta_H0}{estimated null parameter set}

\item{theta_H1}{estimated full parameter set}

\item{other_params}{named list of counts and group assignments}
}
\value{
p-value from two-tailed chi-squared test
}
\description{
Calculate p-value from likelihood ratio test over NBID full and null models
}
\examples{
other_params <- list(x = observed_counts, groups = sample_group_assignments)
theta_H0 <- c(runif(2, min = 0.001, max = 1), runif(1, min = -1, max = 1))
other_params_H0 <- other_params
other_params_H0$null_model <- TRUE
theta_H0_fitted <- optimize.NBID(theta_H0, other_params_H0)
theta_H1 <- c(theta_H0, runif(1, min = -1, max = 1))
other_params_H1 <- other_params
other_params_H1$null_model <- TRUE
theta_H1_fitted <- optimize.NBID(theta_H1, other_params_H1)
pval <- calc_LRT(theta_H0_fitted, theta_H1_fitted, other_params)
}
