% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_ALDEx2}
\alias{call_DA_ALDEx2}
\title{Evaluate differential abundance with ALDEx2}
\usage{
call_DA_ALDEx2(data, groups)
}
\arguments{
\item{data}{simulated data set (samples x features)}

\item{groups}{group (cohort) labels}
}
\value{
unadjusted p-values and "effect sizes" for DA for all features
}
\description{
Evaluate differential abundance with ALDEx2
}
