% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_ANCOMBC}
\alias{call_DA_ANCOMBC}
\title{Evaluate differential abundance with ANCOM-BC}
\usage{
call_DA_ANCOMBC(data, groups)
}
\arguments{
\item{data}{simulated data set}

\item{groups}{group (cohort) labels}

\item{use_TMM}{if TRUE, uses the trimmed mean of M-values normalization}
}
\value{
p-value for DA for all features
}
\description{
Evaluate differential abundance with ANCOM-BC
}
