% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{call_DA_CODA}
\alias{call_DA_CODA}
\title{Fit normal model to logratios and evaluate differential abundance for a focal gene
NOTE: probably need something better than a normal here; it's not a great fit :(
      predict from fitted model for a few of these and see if this is the case}
\usage{
call_DA_CODA(logratios, feature_idx, groups)
}
\arguments{
\item{logratios}{data in ALR format}

\item{feature_idx}{index of gene to test for differential abundance}

\item{groups}{treatment group assignments across samples}
}
\value{
p-value from lm
}
\description{
Fit normal model to logratios and evaluate differential abundance for a focal gene
NOTE: probably need something better than a normal here; it's not a great fit :(
      predict from fitted model for a few of these and see if this is the case
}
