% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{call_DA_counts}
\alias{call_DA_counts}
\title{Fit NBID model to null and full models and evaluate differential abundance for a focal gene; OLD}
\usage{
call_DA_counts(data, feature_idx, call_abundances = TRUE)
}
\arguments{
\item{data}{simulated data set}

\item{feature_idx}{index of gene to test for differential abundance}

\item{call_abundances}{if TRUE, call DA on abundances; if FALSE, on observed counts}
}
\value{
p-value from likelihood ratio test of differential abundance
}
\description{
Fit NBID model to null and full models and evaluate differential abundance for a focal gene; OLD
}
