% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{call_DA_edgeR}
\alias{call_DA_edgeR}
\title{Evaluate differential abundance with edgeR
This evaluates expression on all features of the count matrix together}
\usage{
call_DA_edgeR(data, call_abundances = TRUE, normalization_method = NULL)
}
\arguments{
\item{data}{simulated data set}

\item{call_abundances}{if TRUE, call DA on abundances; if FALSE, on observed counts}

\item{normalization_method}{if NULL, uses library size normalization; other options include "TMM" and "scran"}
}
\value{
p-value associated with group coefficient
}
\description{
Evaluate differential abundance with edgeR
This evaluates expression on all features of the count matrix together
}
