% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_NB}
\alias{call_DA_NB}
\title{Fit negative binomial model to null and full models and evaluate differential abundance for a focal gene}
\usage{
call_DA_NB(data, groups, feature_idx)
}
\arguments{
\item{data}{simulated data set}

\item{groups}{group (cohort) labels}

\item{feature_idx}{index of gene to test for differential abundance}
}
\value{
p-value from NB GLM fit with MASS::glm.nb associated with group coefficient
}
\description{
Fit negative binomial model to null and full models and evaluate differential abundance for a focal gene
}
