% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_Seurat}
\alias{call_DA_Seurat}
\title{Evaluate differential abundance with Wilcox rank sum test or DESeq2 (via Seurat)
This evaluates expression on all features of the count matrix together}
\usage{
call_DA_Seurat(data, groups, method = "DESeq2")
}
\arguments{
\item{data}{simulated data set}

\item{groups}{group (cohort) labels}

\item{method}{options include "wilcox" and "DESeq2"}
}
\value{
p-value for DA for all features
}
\description{
Evaluate differential abundance with Wilcox rank sum test or DESeq2 (via Seurat)
This evaluates expression on all features of the count matrix together
}
