% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{call_DE}
\alias{call_DE}
\title{Fit negative binomial model to null and full models and evaluate differential expression for a focal gene}
\usage{
call_DE(data, feature_idx, call_abundances = TRUE)
}
\arguments{
\item{data}{simulated data set}

\item{feature_idx}{index of gene to test for differential expression}

\item{call_abundances}{if TRUE, call DE on abundances; if FALSE, on observed counts}
}
\value{
p-value from NB GLM fit with MASS::glm.nb
}
\description{
Fit negative binomial model to null and full models and evaluate differential expression for a focal gene
}
