% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_models.R
\name{characterize_dataset}
\alias{characterize_dataset}
\title{Generate simulated differential expression for two conditions}
\usage{
characterize_dataset(counts_A, counts_B)
}
\arguments{
\item{counts_A}{counts associated with the baseline/control condition 
(arranged n x p)}

\item{counts_B}{counts associated with the second condition 
(arranged n x p)}
}
\description{
Generate simulated differential expression for two conditions
}
