% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{compute_sf}
\alias{compute_sf}
\title{Compute reference derived per-sample scale factor}
\usage{
compute_sf(ref_data)
}
\arguments{
\item{ref_data}{count matrix of spike-ins (etc.) in feature x sample 
orientation}
}
\value{
per-sample size factors
}
\description{
Compute reference derived per-sample scale factor
}
