% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{DA_by_edgeR}
\alias{DA_by_edgeR}
\title{Wrapper for call_DA_edgeR generating reference and target calls}
\usage{
DA_by_edgeR(ref_data, data, groups, oracle_calls = NULL, use_TMM = FALSE)
}
\arguments{
\item{ref_data}{absolute abundance data set (samples x features)}

\item{data}{relative abundance data set (samples x features)}

\item{groups}{group (cohort) labels}

\item{oracle_calls}{optional baseline (true) differential abundance calls}

\item{use_TMM}{if TRUE, uses the trimmed mean of M-values normalization}
}
\value{
baseline (true) differential abundance calls and calls made on 
relative abundances
}
\description{
Wrapper for call_DA_edgeR generating reference and target calls
}
