% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{DA_wrapper}
\alias{DA_wrapper}
\title{Wrapper function for differential abundance calling}
\usage{
DA_wrapper(ref_data, data, groups, method = "NBGLM", oracle_calls = NULL)
}
\arguments{
\item{ref_data}{absolute abundance data set (samples x features)}

\item{data}{relative abundance data set (samples x features)}

\item{groups}{group (cohort) labels}

\item{method}{differential abundance calling method (e.g. "DESeq2")}

\item{oracle_calls}{optional baseline (true) differential abundance calls}
}
\value{
named list of baseline (true) differential abundance calls and
calls made on relative abundances (the observed data)
}
\description{
Wrapper function for differential abundance calling
}
