% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_null_distribution}
\alias{generate_null_distribution}
\title{Fit log linear model to null and full models and evaluate differential abundance for a focal gene
This requires a null distribution to have been constructed previously}
\usage{
generate_null_distribution(data, use_abundances = TRUE, n_permutations = 10)
}
\arguments{
\item{data}{simulated data set}

\item{use_abundances}{if TRUE, use abundances to calculate null distro; if FALSE, use observed counts}

\item{n_permutations}{number of times to permute the full data set}
}
\value{
p-value associated with group coefficient
}
\description{
Fit log linear model to null and full models and evaluate differential abundance for a focal gene
This requires a null distribution to have been constructed previously
}
