% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_logratios}
\alias{get_logratios}
\title{Get additive logratios from the data set}
\usage{
get_logratios(data, call_abundances = TRUE)
}
\arguments{
\item{data}{simulated data set}

\item{call_abundances}{if TRUE, call DE on abundances; if FALSE, on observed counts}
}
\value{
logratios
}
\description{
Get additive logratios from the data set
}
