% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate.R
\name{get_timestamp}
\alias{get_timestamp}
\title{Generate a current timestamp in Y-m-d H:M:S format}
\usage{
get_timestamp()
}
\value{
timestamp string
}
\description{
Generate a current timestamp in Y-m-d H:M:S format
}
