% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBID_optimization.R
\name{gradient.NBID}
\alias{gradient.NBID}
\title{Calculate gradients for NBID model}
\usage{
gradient.NBID(theta, other_params)
}
\arguments{
\item{theta}{vector of parameters (see details)}

\item{other_params}{named list of counts, group assignments, size factor (per-sample total counts), and H0 flag}
}
\value{
gradient vector
}
\description{
Calculate gradients for NBID model
}
\details{
Parameter vector contains (in this order) control group dispersion, treatment group dispersion,
focal gene intercept, and (if not evaluating null model) focal gene treatment group effect.
}
\examples{
theta <- c(runif(2, min = 0.001, max = 1), runif(2, min = -1, max = 1))
other_params <- list(x = observed_counts, groups = sample_group_assignments, null_model = FALSE)
grad_vector <- gradient.NBID(theta, other_params)
}
