% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negativeLL_NBID}
\alias{negativeLL_NBID}
\title{Calculate negative log likelihood}
\usage{
negativeLL_NBID(x, counts, groups, size_factors, null_model)
}
\arguments{
\item{x}{parameter vector}

\item{counts}{count vector}

\item{groups}{per-sample group label vector (0 or 1)}

\item{size_factors}{per-sample size factor}

\item{null_model}{boolean indicating whether or not to use a group coefficient}
}
\description{
Calculate negative log likelihood
}
