% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_NBID}
\alias{optim_NBID}
\alias{optimize_NBID}
\title{Optimize the NBID model}
\usage{
optimize_NBID(theta_init, counts, groups, size_factors, null_model, maxit)
}
\arguments{
\item{theta_init}{initial parameter vector}

\item{counts}{count vector}

\item{groups}{per-sample group label vector (0 or 1)}

\item{size_factors}{per-sample size factor}

\item{null_model}{boolean indicating whether or not to use a group coefficient}

\item{maxit}{maximum optimization steps}
}
\description{
Optimize the NBID model
}
