% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_ESCA}
\alias{parse_ESCA}
\title{Parse the TCGA ESCA RNA-seq data}
\usage{
parse_ESCA(absolute = TRUE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the TCGA ESCA RNA-seq data
}
