% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Ferreira}
\alias{parse_Ferreira}
\title{Parse the Ferreira et al. (2014) zebrafish bulk RNA-seq data set}
\usage{
parse_Ferreira(absolute = TRUE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Ferreira et al. (2014) zebrafish bulk RNA-seq data set
}
