% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Gruen}
\alias{parse_Gruen}
\title{Parse the Gruen et al. (2014) single-cell cell line data set}
\usage{
parse_Gruen(absolute = TRUE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Gruen et al. (2014) single-cell cell line data set
}
