% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Hagai}
\alias{parse_Hagai}
\title{Parse the Hagai et al. (2018) cross-species, cross-cell, multiple condition
experiment for mouse unstimulated vs. pIC4 time point}
\usage{
parse_Hagai(absolute = TRUE, use_cpm = FALSE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Hagai et al. (2018) cross-species, cross-cell, multiple condition
experiment for mouse unstimulated vs. pIC4 time point
}
