% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Hashimshony}
\alias{parse_Hashimshony}
\title{Parse the Hashimshony et al. (2016) mouse fibroblast data set}
\usage{
parse_Hashimshony(absolute = TRUE, use_cpm = FALSE, hkg_list = NULL)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}

\item{use_cpm}{flag indicating whether or not to convert observed counts to
counts per million (CPM)}

\item{hkg_list}{optional housekeeping gene list; test spike-in abundances
against the abundances of these genes, if present}
}
\value{
named list of counts and group labels
}
\description{
Parse the Hashimshony et al. (2016) mouse fibroblast data set
}
