% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Kimmerling}
\alias{parse_Kimmerling}
\title{Parse the Kimmerling et al. (2018) fibroblast data set}
\usage{
parse_Kimmerling(
  absolute = TRUE,
  use_spike_ins = FALSE,
  use_cpm = FALSE,
  hkg_list = NULL
)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}

\item{use_cpm}{flag indicating whether or not to convert observed counts to
counts per million (CPM)}

\item{hkg_list}{dummy list of control features (not used)}
}
\value{
named list of counts and group labels
}
\description{
Parse the Kimmerling et al. (2018) fibroblast data set
}
