% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Klein}
\alias{parse_Klein}
\title{Parse the Klein et al. (2015) mouse +/- LIF data set}
\usage{
parse_Klein(absolute = TRUE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Klein et al. (2015) mouse +/- LIF data set
}
