% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Morton}
\alias{parse_Morton}
\title{Parse the Morton et al. (2019) 16S data}
\usage{
parse_Morton()
}
\value{
named list of counts and group labels
}
\description{
Parse the Morton et al. (2019) 16S data
}
