% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Song}
\alias{parse_Song}
\title{Parse the Song et al. (2021) nCounter data}
\usage{
parse_Song(absolute = TRUE, use_cpm = FALSE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Song et al. (2021) nCounter data
}
