% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate.R
\name{plot_bipartite_graph}
\alias{plot_bipartite_graph}
\title{Generate a bipartite graph for differential abundance/expression}
\usage{
plot_bipartite_graph(expr1, expr2, alpha = 1)
}
\arguments{
\item{expr1}{expression vector in baseline condition}

\item{expr2}{expression vector in differential condition}

\item{alpha}{optional transparency parameter}
}
\description{
Generate a bipartite graph for differential abundance/expression
}
