% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilerplate.R
\name{plot_stacked_bars}
\alias{plot_stacked_bars}
\title{Generate a color palette suitable for stacked bar plots given a feature 
number (S)}
\usage{
plot_stacked_bars(data, palette = NULL, save_name = NULL)
}
\arguments{
\item{data}{matrix of abundances or observed counts (samples x features)}

\item{palette}{a color palette}

\item{save_name}{optional name to save plot under}
}
\description{
Generate a color palette suitable for stacked bar plots given a feature 
number (S)
}
