% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBID_optimization.R
\name{predict.NBID}
\alias{predict.NBID}
\title{Attempt to recreate the observed counts for a given gene from the fitted model (giving a means to eyeball
model goodness-of-fit)}
\usage{
predict.NBID(theta, other_params)
}
\arguments{
\item{theta}{vector of parameters (see details)}

\item{other_params}{named list of counts, group assignments, size factor (per-sample total counts), and H0 flag}
}
\value{
a set of predicted counts from the fitted model
}
\description{
Attempt to recreate the observed counts for a given gene from the fitted model (giving a means to eyeball
model goodness-of-fit)
}
\details{
Parameter vector contains (in this order) control group dispersion, treatment group dispersion,
focal gene intercept, and (if not evaluating null model) focal gene treatment group effect.
}
