% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R, R/simulate.R
\name{record_result}
\alias{record_result}
\title{Write simulation result to a file}
\usage{
record_result(out_str, out_file)

record_result(out_str, out_file)
}
\arguments{
\item{out_str}{string to write to file}

\item{out_file}{log file path and name}
}
\description{
Write simulation result to a file

Write simulation result to a file
}
