% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{resample_counts}
\alias{resample_counts}
\title{Utility function to multinomially resample counts}
\usage{
resample_counts(proportions, library_sizes)
}
\arguments{
\item{proportions}{relative abundance table (samples x features)}
}
\value{
resampled count table
}
\description{
Utility function to multinomially resample counts
}
