% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_evaluation_instance}
\alias{run_evaluation_instance}
\title{Generate simulated data with randomly perturbed features, evaluate false negatives and positives in
differential expression calls and write these to an output file}
\usage{
run_evaluation_instance(
  p,
  fold_change,
  proportion_perturbed_features,
  run_label,
  mean_size_factor = 5000,
  size_factor_correlation = 0,
  bimodal = FALSE,
  output_file = "results.txt",
  alpha = 0.05
)
}
\arguments{
\item{p}{number of genes}

\item{fold_change}{fold change to induce in perturbed features}

\item{proportion_perturbed_features}{proportion of features to differentially express}

\item{run_label}{human-readable string identifier for this simulation run}

\item{mean_size_factor}{average sequencing depth (total counts per sample)}

\item{size_factor_correlation}{correlation of observed abundances to original total abundances}

\item{bimodal}{if TRUE, simulates a composition with a high and low abundance cohorts}

\item{output_file}{output file to append to}

\item{alpha}{significant level below which to call a feature differentially expressed}
}
\description{
Generate simulated data with randomly perturbed features, evaluate false negatives and positives in
differential expression calls and write these to an output file
}
\details{
Writes out error and simulation run statistics to a file.
}
