% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{run_RNAseq_evaluation_instance}
\alias{run_RNAseq_evaluation_instance}
\title{Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate error in
differential abundance calls and write these to an output file}
\usage{
run_RNAseq_evaluation_instance(
  p,
  n,
  proportion_da,
  k = 1,
  library_size_correlation = 0,
  alpha = 0.05,
  use_ALR = FALSE,
  filter_abundance = 1,
  methods = c("NB", "edgeR")
)
}
\arguments{
\item{p}{number of genes}

\item{n}{number of samples per group}

\item{proportion_da}{proportion of differentially abundant genes}

\item{k}{number of cell types to simulate}

\item{library_size_correlation}{correlation of observed abundances to original total abundances}

\item{alpha}{significant level below which to call a feature differentially abundant}

\item{use_ALR}{if TRUE, evaluates differential abundance using a spike-in and the additive logratio}

\item{filter_abundance}{the minimum average abundance to require of features we'll evaluate for differential abundance
(e.g. a filter_abundance of 1 evaluates features with average abundance across conditions of at least 1)}

\item{methods}{vector of differential abundance testing methods to use; options are "NB", "GLM", "edgeR"}
}
\description{
Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate error in
differential abundance calls and write these to an output file
}
\details{
Writes out error and simulation run statistics to a file.
}
