% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_models.R
\name{scaled_counts_scran}
\alias{scaled_counts_scran}
\title{Scale counts using scran-like size factors}
\usage{
scaled_counts_scran(counts, groups, pseudocount = 0.5)
}
\arguments{
\item{counts}{count table oriented as samples (rows) x features (columns)}

\item{groups}{group or treatment labels for columns}

\item{pseudocount}{small count to augment whole matrix by in order to 
eliminate zero counts}
}
\value{
scaled count table oriented as samples (rows) x features (columns)
}
\description{
Scale counts using scran-like size factors
}
