% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate counts for control vs. treatment groups}
\usage{
simulate_data(
  p,
  n,
  proportions,
  perturbed_features,
  fold_change = NULL,
  mean_size_factor = 5000,
  size_factor_correlation = 0
)
}
\arguments{
\item{p}{number of genes}

\item{n}{number of samples per group}

\item{proportions}{set of proportions from which to sample true abundances}

\item{perturbed_features}{number of features to differentially express}

\item{fold_change}{fold change to induce in perturbed features}

\item{mean_size_factor}{average total abundance to simulate}

\item{size_factor_correlation}{correlation of observed abundances to original total abundances}
}
\value{
named list of true abundances, observed abundances, and group assignments
}
\description{
Simulate counts for control vs. treatment groups
}
