% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_sequence_counts}
\alias{simulate_sequence_counts}
\title{Simulate UMI count single-cell RNA-seq data}
\usage{
simulate_sequence_counts(
  n = 500,
  p = 1000,
  ref_data_name = "Athanasidou_ciona",
  data_obj = NULL,
  asymmetry = 0.5,
  proportion_da = 0.1,
  spike_in = FALSE
)
}
\arguments{
\item{n}{number of samples per group}

\item{p}{number of features (i.e. genes or bacterial taxa)}

\item{ref_data_name}{optional: specifies the reference data set to use
("Morton", "Athanasiadou_ciona", "simulated"; case-sensitive)}

\item{data_obj}{reference data set object; if non-null, this takes precedence
over ref_data_name}

\item{asymmetry}{proportion of baselines to draw from condition 1}

\item{proportion_da}{proportion of differentially abundant genes}

\item{spike_in}{simulate a control spike in with low dispersion}
}
\value{
named list of true abundances, observed abundances, and group assignments
}
\description{
Simulate UMI count single-cell RNA-seq data
}
\details{
Library size correlation can be specified as 0, 1, or 2, indicating no correlation, modest correlation, or very strong
correlation with true abundances.

Samples the mean gene abundances from Halpern et al. (2017) as a references. This gives a data set
with on average 75% zeros which may still not be representative in terms of sparsity. Need to follow up. (7/22/2020)
}
