% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spike_in_ones}
\alias{spike_in_ones}
\title{Utility function to spike-in a one-count to prevent any fully unobserved taxa}
\usage{
spike_in_ones(counts)
}
\arguments{
\item{counts}{count table (samples x features)}
}
\value{
counts with a minimum single observation of each feature
}
\description{
Utility function to spike-in a one-count to prevent any fully unobserved taxa
}
