% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sweep_RNAseq}
\alias{sweep_RNAseq}
\title{Generate simulated data with RNA-seq like features and > 3 fold differential expression, evaluate false negatives and positives in
differential expression calls and write these to an output file}
\usage{
sweep_RNAseq(p, n, run_label, de_sweep = seq(from = 0.1, to = 0.9, by =
  0.1), corr_sweep = seq(from = 0.1, to = 0.9, by = 0.1),
  output_file = "results.txt", alpha = 0.05, use_ALR = FALSE,
  rarefy = FALSE)
}
\arguments{
\item{p}{number of genes}

\item{n}{number of samples per group}

\item{run_label}{human-readable string identifier for this simulation run}

\item{de_sweep}{vector of proportions of differentially expressed genes to simulate}

\item{corr_sweep}{vector of correlations of size factors (true vs. observed abundances) to simulate}

\item{output_file}{output file to append to}

\item{alpha}{significant level below which to call a feature differentially expressed}

\item{use_ALR}{if TRUE, evaluates differential expression using a spike-in and the additive logratio}

\item{rarefy}{if TRUE, resamples the counts in each sample to the lowest observed total counts}
}
\description{
Generate simulated data with RNA-seq like features and > 3 fold differential expression, evaluate false negatives and positives in
differential expression calls and write these to an output file
}
\details{
Writes out error and simulation run statistics to a file.
}
