% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{sweep_simulations}
\alias{sweep_simulations}
\title{Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate false negatives and positives in
differential abundance calls and write these to an output file}
\usage{
sweep_simulations(
  p,
  n,
  k = 1,
  de_sweep = seq(from = 0.1, to = 0.9, by = 0.1),
  corr_sweep = seq(from = 0.1, to = 0.9, by = 0.1),
  alpha = 0.05,
  use_ALR = FALSE,
  filter_abundance = 0,
  methods = c("NB", "edgeR")
)
}
\arguments{
\item{p}{number of genes}

\item{n}{number of samples per group}

\item{k}{number of cell types to simulate}

\item{de_sweep}{vector of proportions of differentially abundant genes to simulate}

\item{corr_sweep}{vector of correlations of size factors (true vs. observed abundances) to simulate}

\item{alpha}{significant level below which to call a feature differentially abundant}

\item{use_ALR}{if TRUE, evaluates differential abundance using a spike-in and the additive logratio}

\item{filter_abundance}{the minimum average abundance to require of features we'll evaluate for differential abundance
(e.g. a filter_abundance of 1 evaluates features with average abundance across conditions of at least 1)}

\item{methods}{vector of differential abundance testing methods to use; options are "NB", "GLM", "edgeR"}
}
\description{
Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate false negatives and positives in
differential abundance calls and write these to an output file
}
\details{
Writes out error and simulation run statistics to a file.
}
