% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{update_metadata}
\alias{update_metadata}
\title{#' Write simulation result to a file
#'
#' @param out_str string to write to file
#' @param out_file log file path and name
#' @return NULL
#' @import filelock
#' @export
record_result <- function(out_str, out_file) {
  lock_file <- paste0(out_file, ".lck")
  lck <- lock(lock_file, timeout = Inf)
  write(out_str, file = out_file, append = TRUE)
  unlock(lck)
}
Update the metadata file in the simulated data set directory}
\usage{
update_metadata(filename, p, n, proportion_da, size_factor_correlation, k)
}
\arguments{
\item{filename}{autogenerated filename for simulated data set}

\item{p}{number of genes}

\item{n}{number of samples per group}

\item{proportion_da}{proportion of differentially abundant genes}

\item{size_factor_correlation}{correlation of observed abundances to original total abundances}

\item{k}{number of cell types to simulate; if not NULL, single-cell data is simulated; if NULL, bulk RNA-seq data is simulated}
}
\description{
#' Write simulation result to a file
#'
#' @param out_str string to write to file
#' @param out_file log file path and name
#' @return NULL
#' @import filelock
#' @export
record_result <- function(out_str, out_file) {
  lock_file <- paste0(out_file, ".lck")
  lck <- lock(lock_file, timeout = Inf)
  write(out_str, file = out_file, append = TRUE)
  unlock(lck)
}
Update the metadata file in the simulated data set directory
}
