% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{wrangle_validation_data}
\alias{wrangle_validation_data}
\title{Compute reference-derived per-sample scale factor}
\usage{
wrangle_validation_data(
  dataset_name,
  threshold,
  use_cpm,
  testing = FALSE,
  hkg_list = NULL
)
}
\arguments{
\item{dataset_name}{short name of validation data set to use (e.g. Barlow)}

\item{threshold}{mean count lower limit for feature inclusion}

\item{use_cpm}{flag indicating whether or not to convert observed counts to
counts per million (CPM)}

\item{testing}{flag indicating whether or not to subsample for (quick) testing}

\item{hkg_list}{optional housekeeping gene list; test spike-in abundances
against the abundances of these genes, if present}
}
\value{
absolute, relative (observed) abundances and feature names after 
filtering out very low abundance features
}
\description{
Compute reference-derived per-sample scale factor
}
