% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_references.R
\name{build_Barlow_reference}
\alias{build_Barlow_reference}
\title{Generate differential expression reference for Barlow et al. (2020) 16S data}
\usage{
build_Barlow_reference()
}
\value{
named list of observed counts in conditions 1 and 2
}
\description{
Generate differential expression reference for Barlow et al. (2020) 16S data
}
