% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_references.R
\name{build_Morton_reference}
\alias{build_Morton_reference}
\title{Generate differential expression reference for Morton et al. (2019) 16S data}
\usage{
build_Morton_reference()
}
\description{
Generate differential expression reference for Morton et al. (2019) 16S data
}
