% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_references.R
\name{build_simulated_reference}
\alias{build_simulated_reference}
\title{Generate simulated differential expression for two conditions}
\usage{
build_simulated_reference(
  p = 1000,
  log_mean = 1,
  log_var = 4,
  perturb_size = 0.5,
  base_correlation = NULL,
  concentration = 1e+06,
  save_name = NULL
)
}
\arguments{
\item{p}{number of features (genes, taxa) to simulate}

\item{log_mean}{log mean for condition 1}

\item{log_var}{log variance for condition 1}

\item{perturb_size}{percent of the log mean to use (on average) as the size 
of the noise component added to condition 1 to give condition 2}

\item{base_correlation}{this is the base correlation matrix for simulated
features in log space}

\item{concentration}{concentration parameter for sampling the correlation 
across feature; if base_correlation is NULL and concentration = LARGE, features
are effectively independent}

\item{save_name}{optional tag to append to saved data file}
}
\description{
Generate simulated differential expression for two conditions
}
