% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_references.R
\name{build_simulated_reference}
\alias{build_simulated_reference}
\title{Generate simulated differential expression for two conditions}
\usage{
build_simulated_reference(
  p = 1000,
  log_mean = 0,
  log_var = 2,
  log_noise_var = 1,
  base_correlation = 0,
  concentration = 1e+06,
  save_name = NULL
)
}
\arguments{
\item{p}{number of features (genes, taxa) to simulate}

\item{log_mean}{log mean for condition 1}

\item{log_var}{log variance for condition 1}

\item{log_noise_var}{log variance for noise component added to condition 1
to give condition 2}

\item{base_correlation}{this is the base correlation (in log space) simulated
across features}

\item{concentration}{concentration parameter for sampling the correlation 
across feature; if base_correlation = 0 and concentration = LARGE, features
are effectively independent}

\item{save_name}{optional tag to append to saved data file}
}
\description{
Generate simulated differential expression for two conditions
}
