% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{calc_DA_discrepancy}
\alias{calc_DA_discrepancy}
\title{Calculates the discrepancy between differential abundance calls made on
absolute and relative abundances}
\usage{
calc_DA_discrepancy(
  calls,
  oracle_calls,
  adjusted = TRUE,
  alpha = 0.05,
  beta = NULL
)
}
\arguments{
\item{calls}{calls (as a numeric vector of p-values or a character string 
that can be parsed into one) made on relative abundances (the observed data)}

\item{oracle_calls}{calls (as a numeric vector of p-values or a character 
string that can be parsed into one) made on the absolute abundances}

\item{adjusted}{flag indicating whether or not to perform multiple test
correction}

\item{alpha}{significance threshold after FD adjustment}

\item{beta}{optional effect size threshold at the scale of the original 
abundances (i.e. not log scale)}
}
\value{
named list of all calls, TPR, and FPR
}
\description{
Calculates the discrepancy between differential abundance calls made on
absolute and relative abundances
}
