% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{calc_threshold_DA}
\alias{calc_threshold_DA}
\title{Threshold to generate calls on differentially abundant features}
\usage{
calc_threshold_DA(counts, nA = NULL)
}
\arguments{
\item{counts}{abundance data set (samples x features)}

\item{nA}{optional parameter specifying the number of samples in condition A}
}
\value{
differential calls in the form of faux p-values (0 for differential,
1 for not differential)
}
\description{
Threshold to generate calls on differentially abundant features
}
