% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_ALDEx2}
\alias{call_DA_ALDEx2}
\title{Evaluate differential abundance with Welch's t-test using ALDEx2 (w/ iqlr)
This evaluates expression on all features of the count matrix together}
\usage{
call_DA_ALDEx2(data, groups)
}
\arguments{
\item{data}{simulated data set}

\item{groups}{group (cohort) labels}
}
\value{
p-value for DA for all features
}
\description{
Evaluate differential abundance with Welch's t-test using ALDEx2 (w/ iqlr)
This evaluates expression on all features of the count matrix together
}
