% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_edgeR}
\alias{call_DA_edgeR}
\title{Evaluate differential abundance with edgeR}
\usage{
call_DA_edgeR(data, groups, use_TMM = FALSE)
}
\arguments{
\item{data}{simulated data set}

\item{groups}{group (cohort) labels}

\item{use_TMM}{if TRUE, uses the trimmed mean of M-values normalization}
}
\value{
unadjusted p-values and "effect sizes" for DA for all features
}
\description{
Evaluate differential abundance with edgeR
}
