% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_LM}
\alias{call_DA_LM}
\title{Fit log linear model to null and full models and evaluate differential abundance for a focal gene}
\usage{
call_DA_LM(data, feature_idx, call_abundances = TRUE)
}
\arguments{
\item{data}{simulated data set}

\item{feature_idx}{index of gene to test for differential abundance}

\item{call_abundances}{if TRUE, call DA on abundances; if FALSE, on observed counts}
}
\value{
p-value associated with group coefficient
}
\description{
Fit log linear model to null and full models and evaluate differential abundance for a focal gene
}
