% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_MAST}
\alias{call_DA_MAST}
\title{Evaluate differential abundance with Wilcox rank sum test or DESeq2 (via Seurat)
This evaluates expression on all features of the count matrix together}
\usage{
call_DA_MAST(data, groups)
}
\arguments{
\item{data}{simulated data set}

\item{groups}{group (cohort) labels}
}
\value{
p-value for DA for all features
}
\description{
Evaluate differential abundance with Wilcox rank sum test or DESeq2 (via Seurat)
This evaluates expression on all features of the count matrix together
}
