% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{call_DA_NB}
\alias{call_DA_NB}
\title{Fit negative binomial model to null and full models and evaluate differential abundance for a focal gene}
\usage{
call_DA_NB(data, feature_idx, call_abundances = TRUE, rarefy_depth = 0)
}
\arguments{
\item{data}{simulated data set}

\item{feature_idx}{index of gene to test for differential abundance}

\item{call_abundances}{if TRUE, call DA on abundances; if FALSE, on observed counts}

\item{rarefy_depth}{if non-zero, specifies the depth at which to resample the counts}
}
\value{
p-value from NB GLM fit with MASS::glm.nb associated with group coefficient
}
\description{
Fit negative binomial model to null and full models and evaluate differential abundance for a focal gene
}
