% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{call_DA_scran}
\alias{call_DA_scran}
\title{Apply scran normalization and call differentially expressed (marker) genes}
\usage{
call_DA_scran(data, groups)
}
\arguments{
\item{data}{simulated data set}

\item{groups}{group (cohort) labels}
}
\value{
binary calls for all genes as pseudo-pvalues (0 = DE, 1 = not DE)
}
\description{
Apply scran normalization and call differentially expressed (marker) genes
}
