% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R
\name{evaluate_DA}
\alias{evaluate_DA}
\title{Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate error in
differential abundance calls and write these to an output file}
\usage{
evaluate_DA(
  data,
  alpha = 0.05,
  use_ALR = FALSE,
  filter_abundance = 0,
  method = "edgeR"
)
}
\arguments{
\item{data}{simulated data set}

\item{alpha}{significant level below which to call a feature differentially abundant}

\item{use_ALR}{if TRUE, evaluates differential abundance using a spike-in and the additive logratio}

\item{filter_abundance}{the minimum average abundance to require of features we'll evaluate for differential abundance
(e.g. a filter_abundance of 1 evaluates features with average abundance across conditions of at least 1)}

\item{method}{differential abundance testing method to use; options are "NB", "GLM", "edgeR", "edgeR_TMM", "edgeR_scran"}
}
\description{
Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate error in
differential abundance calls and write these to an output file
}
\details{
Writes out error and simulation run statistics to a file.
}
