% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_models.R
\name{fit_predictive_model}
\alias{fit_predictive_model}
\title{Generate simulated differential expression for two conditions}
\usage{
fit_predictive_model(
  model = "linear",
  do_predict = FALSE,
  DE_method = "all",
  plot_weights = FALSE
)
}
\arguments{
\item{model}{options are "linear", "RF" (random forest), "EN" (elastic net),
"GP" (Gaussian process)}

\item{do_predict}{flag indicating whether or not to generated predictions on a
test set}

\item{DE_method}{which DE calling method's results to predict; if "all",
prediction is over all results together}

\item{plot_weights}{flag indicating whether or not to plot some visualization
of feature weights}
}
\value{
fitted model
}
\description{
Generate simulated differential expression for two conditions
}
