% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_models.R
\name{fit_predictive_model}
\alias{fit_predictive_model}
\title{Generate simulated differential expression for two conditions}
\usage{
fit_predictive_model(
  DE_methods = c("ALDEx2", "DESeq2", "scran"),
  use_baseline = "oracle",
  use_totals = FALSE,
  use_renorm_counts = FALSE,
  output_weights = TRUE,
  train_percent = 0.8
)
}
\arguments{
\item{DE_methods}{which DE calling method's results to predict; if "all",
prediction is over all results together}

\item{use_baseline}{"self" or "oracle"}

\item{use_totals}{pull fold change in total estimate from absolute counts}

\item{use_renorm_counts}{pull features generated from method-renormalized
data}

\item{output_weights}{flag indicating whether or not to plot some visualization
of feature weights}

\item{train_percent}{percent of simulated datasets to train on}
}
\value{
NULL (fitted models are saved in output directory)
}
\description{
Generate simulated differential expression for two conditions
}
