% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_models.R
\name{fit_predictive_model}
\alias{fit_predictive_model}
\title{Generate simulated differential expression for two conditions}
\usage{
fit_predictive_model(
  DE_method = "all",
  plot_weights = FALSE,
  exclude_partials = TRUE,
  exclude_independent = FALSE,
  train_percent = 0.8,
  save_tag = ""
)
}
\arguments{
\item{DE_method}{which DE calling method's results to predict; if "all",
prediction is over all results together}

\item{plot_weights}{flag indicating whether or not to plot some visualization
of feature weights}

\item{exclude_partials}{flag indicating whether to exclude simulations with 
partial abundance information}

\item{exclude_independent}{flag indicating whether to exclude simulations 
with uncorrelated features}

\item{train_percent}{percent of simulated datasets to train on}

\item{save_tag}{optional tag to append to file name for saved predictive model}
}
\value{
NULL (fitted models are saved in output directory)
}
\description{
Generate simulated differential expression for two conditions
}
