% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{normalize_Athanasiadou}
\alias{normalize_Athanasiadou}
\title{Utility function that applies the ERCC normalization code from Athanasiadou
et al. (2021)}
\usage{
normalize_Athanasiadou(file_dir, SI_file, RNA_file, ERCC_annot, groups)
}
\arguments{
\item{file_dir}{base file directory}

\item{SI_file}{experiment SI file}

\item{RNA_file}{experiment RNA file}

\item{ERCC_annot}{experiment ERCC file}

\item{groups}{per-sample group labels}
}
\value{
matrix of rescaled count data
}
\description{
Utility function that applies the ERCC normalization code from Athanasiadou
et al. (2021)
}
