% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Athanasiadou}
\alias{parse_Athanasiadou}
\title{Parse the Athanasiadou et al. (2021) bulk RNA-seq data}
\usage{
parse_Athanasiadou(which_data = "ciona")
}
\arguments{
\item{which_data}{options are "ciona" or "yeast"}
}
\value{
named list of counts and group labels
}
\description{
Parse the Athanasiadou et al. (2021) bulk RNA-seq data
}
