% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Barlow}
\alias{parse_Barlow}
\title{Parse the Barlow et al. (2020) 16S data}
\usage{
parse_Barlow()
}
\value{
named list of counts and group labels
}
\description{
Parse the Barlow et al. (2020) 16S data
}
