% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Hagai}
\alias{parse_Hagai}
\title{Parse the Hagai et al. (2018) cross-species, cross-cell, multiple condition
experiment for mouse unstimulated vs. pIC4 time point}
\usage{
parse_Hagai(absolute = TRUE, use_cpm = FALSE, hkg_list = NULL)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}

\item{use_cpm}{flag indicating whether or not to convert observed counts to
counts per million (CPM)}

\item{hkg_list}{optional housekeeping gene list; test spike-in abundances
against the abundances of these genes, if present}
}
\value{
named list of counts and group labels
}
\description{
Parse the Hagai et al. (2018) cross-species, cross-cell, multiple condition
experiment for mouse unstimulated vs. pIC4 time point
}
