% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Hashimshony}
\alias{parse_Hashimshony}
\title{Parse the Hashimshony et al. (2016) mouse fibroblast data set}
\usage{
parse_Hashimshony(absolute = TRUE, use_cpm = FALSE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Hashimshony et al. (2016) mouse fibroblast data set
}
