% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Kimmerling}
\alias{parse_Kimmerling}
\title{Parse the Kimmerling et al. (2018) fibroblast data set}
\usage{
parse_Kimmerling(absolute = TRUE, use_spike_ins = FALSE, use_cpm = FALSE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Kimmerling et al. (2018) fibroblast data set
}
