% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Monaco}
\alias{parse_Monaco}
\title{Parse the Monaco et al. (2019) immune cell data set}
\usage{
parse_Monaco(absolute = TRUE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Monaco et al. (2019) immune cell data set
}
