% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Muraro}
\alias{parse_Muraro}
\title{Parse the Muraro et al. (2016) single-cell pancreas data set}
\usage{
parse_Muraro(absolute = TRUE, use_cpm = FALSE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Muraro et al. (2016) single-cell pancreas data set
}
