% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Owens}
\alias{parse_Owens}
\title{Parse the Owens et al. (2016) zebrafish developmental series}
\usage{
parse_Owens(absolute = TRUE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}
}
\value{
named list of counts and group labels
}
\description{
Parse the Owens et al. (2016) zebrafish developmental series
}
