% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_VieiraSilva}
\alias{parse_VieiraSilva}
\title{Parse the Vieira-Silva et al. (2019) QMP data set}
\usage{
parse_VieiraSilva(absolute = TRUE, use_cpm = FALSE, hkg_list = NULL)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}

\item{use_cpm}{flag indicating whether or not to convert observed counts to
counts per million (CPM)}

\item{hkg_list}{dummy argument for control features (not used)}
}
\value{
named list of counts and group labels
}
\description{
Parse the Vieira-Silva et al. (2019) QMP data set
}
