% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasets.R
\name{parse_Yu}
\alias{parse_Yu}
\title{Parse the Yu et al. (2014) rat tissue bulk RNA-seq}
\usage{
parse_Yu(absolute = TRUE, use_cpm = FALSE)
}
\arguments{
\item{absolute}{flag indicating whether or not to parse absolute abundance
data}

\item{use_cpm}{flag indicating whether or not to convert observed counts to
counts per million (CPM)}
}
\value{
named list of counts and group labels
}
\description{
Parse the Yu et al. (2014) rat tissue bulk RNA-seq
}
