% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_models.R
\name{pull_features}
\alias{pull_features}
\title{Generate simulated differential expression for two conditions}
\usage{
pull_features(
  DE_methods = c("ALDEx2", "DESeq2", "scran"),
  use_baseline = "oracle",
  use_totals = FALSE,
  use_renorm_counts = FALSE,
  use_cpm = FALSE
)
}
\arguments{
\item{DE_methods}{differential abundance calling methods to use}

\item{use_baseline}{one of either "self" or "oracle"; baseline differential
abundance calls against which we will score accuracy of calls made on 
observed (relative) abundance data}

\item{use_totals}{pull fold change in total estimate from absolute counts}

\item{use_renorm_counts}{pull features generated from method-renormalized
data}

\item{use_cpm}{use counts per million for relative abundances}
}
\value{
data.frame with predictive model training features
}
\description{
Generate simulated differential expression for two conditions
}
