% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pred_models.R
\name{pull_features}
\alias{pull_features}
\title{Generate simulated differential expression for two conditions}
\usage{
pull_features(
  DE_methods = c("ALDEx2", "DESeq2", "scran"),
  use_baseline = "self",
  exclude_partials = TRUE,
  exclude_independent = FALSE,
  fit_full = FALSE,
  feature_list = NULL
)
}
\arguments{
\item{exclude_partials}{flag indicating whether to exclude simulations with 
partial abundance information}

\item{exclude_independent}{flag indicating whether to exclude simulations 
with uncorrelated features}

\item{fit_full}{fit model with true fold change information}
}
\value{
data.frame with predictive model training features
}
\description{
Generate simulated differential expression for two conditions
}
