% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{resample_counts}
\alias{resample_counts}
\title{Simulate sequence counts according to a negative binomial}
\usage{
resample_counts(theta, groups)
}
\arguments{
\item{theta}{vector of (2) group dispersion parameters and (2) per-gene log covariates is theta}

\item{groups}{condition (e.g. control vs. treatment) labels}
}
\value{
named list of true abundances and observed abundances
}
\description{
Simulate sequence counts according to a negative binomial
}
