% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{resample_counts}
\alias{resample_counts}
\title{Simulate sequence counts according to a negative binomial}
\usage{
resample_counts(theta, groups, library_size_correlation, spike_in = FALSE)
}
\arguments{
\item{theta}{vector of (2) group dispersion parameters and (2) per-gene log covariates is theta}

\item{groups}{condition (e.g. control vs. treatment) labels}

\item{library_size_correlation}{correlation of observed abundances to original absolute abundances}

\item{spike_in}{flag indicating whether or not to simulate a spike-in (control) with very low dispersion}
}
\value{
named list of true abundances and observed abundances
}
\description{
Simulate sequence counts according to a negative binomial
}
