% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_DA.R, R/simulate.R
\name{run_RNAseq_evaluation_instance}
\alias{run_RNAseq_evaluation_instance}
\title{Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate error in
differential abundance calls and write these to an output file}
\usage{
run_RNAseq_evaluation_instance(
  p,
  n,
  proportion_da,
  k = NULL,
  size_factor_correlation = 0,
  output_file = "results.txt",
  alpha = 0.05,
  use_ALR = FALSE,
  filter_abundance = 1,
  call_DA_by_NB = TRUE,
  rarefy = FALSE
)

run_RNAseq_evaluation_instance(
  p,
  n,
  proportion_da,
  k = NULL,
  size_factor_correlation = 0,
  output_file = "results.txt",
  alpha = 0.05,
  use_ALR = FALSE,
  filter_abundance = 1,
  call_DA_by_NB = TRUE,
  rarefy = FALSE
)
}
\arguments{
\item{p}{number of genes}

\item{n}{number of samples per group}

\item{proportion_da}{proportion of differentially abundant genes}

\item{k}{number of cell types to simulate; if not NULL, single-cell data is simulated; if NULL, bulk RNA-seq data is simulated}

\item{size_factor_correlation}{correlation of observed abundances to original total abundances}

\item{output_file}{output file to append to}

\item{alpha}{significant level below which to call a feature differentially abundant}

\item{use_ALR}{if TRUE, evaluates differential abundance using a spike-in and the additive logratio}

\item{filter_abundance}{the minimum average abundance to require of features we'll evaluate for differential abundance
(e.g. a filter_abundance of 1 evaluates features with average abundance across conditions of at least 1)}

\item{call_DA_by_NB}{if TRUE, uses a NB GLM to call differential abundance; if FALSE, uses a log linear model with a
permutation test}

\item{rarefy}{if TRUE, resamples the counts in each sample to the lowest observed total counts}

\item{run_label}{human-readable string identifier for this simulation run}
}
\description{
Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate error in
differential abundance calls and write these to an output file

Generate simulated data with RNA-seq like features and > 3 fold differential abundance, evaluate error in
differential abundance calls and write these to an output file
}
\details{
Writes out error and simulation run statistics to a file.

Writes out error and simulation run statistics to a file.
}
