% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_RNAseq}
\alias{simulate_RNAseq}
\title{Simulate RNA-seq (like) data}
\usage{
simulate_RNAseq(
  p = 5000,
  n = 500,
  proportion_de = 0.5,
  size_factor_correlation = 0
)
}
\arguments{
\item{p}{number of genes}

\item{n}{number of samples per group}

\item{proportion_de}{proportion of differentially expressed genes}

\item{size_factor_correlation}{correlation of observed abundances to original total abundances}
}
\value{
named list of true abundances, observed abundances, and group assignments
}
\description{
Simulate RNA-seq (like) data
}
