% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_sequence_counts}
\alias{simulate_sequence_counts}
\title{Simulate sequence count data}
\usage{
simulate_sequence_counts(
  baseline_log_abundances,
  n = 500,
  proportion_de = 0.5,
  size_factor_correlation = 0,
  spike_in = FALSE
)
}
\arguments{
\item{baseline_log_abundances}{a distribution (length p) of baseline log abundances from which to simulate}

\item{n}{number of samples per group}

\item{proportion_de}{proportion of differentially abundant genes}

\item{size_factor_correlation}{correlation of observed abundances to original total abundances}

\item{spike_in}{simulate a control spike in with low dispersion}
}
\value{
named list of true abundances, observed abundances, and group assignments
}
\description{
Simulate sequence count data
}
