% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_sequence_counts}
\alias{simulate_sequence_counts}
\title{Simulate UMI count single-cell RNA-seq data}
\usage{
simulate_sequence_counts(
  n = 20,
  p = 100,
  ref_data_name = "simulated",
  data_obj = NULL,
  replicate_noise = 0,
  proportion_da = 0.75
)
}
\arguments{
\item{n}{number of samples per group}

\item{p}{number of features (i.e. genes or bacterial taxa)}

\item{ref_data_name}{optional name of reference data set to use from among
"Morton", "Athanasiadou_ciona", "simulated"}

\item{data_obj}{optional reference data set object; this takes precedence 
over ref_data_name}

\item{replicate_noise}{optional deviation parameter associated with variation
in replicate totals (a max of 1 is approx. an upper limit)}

\item{proportion_da}{optional proportion of differentially abundant genes}
}
\value{
named list of true abundances, observed abundances, and group 
assignments
}
\description{
Simulate UMI count single-cell RNA-seq data
}
