% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_singlecell_RNAseq}
\alias{simulate_singlecell_RNAseq}
\title{Simulate UMI count single-cell RNA-seq data}
\usage{
simulate_singlecell_RNAseq(
  p = 20000,
  n = 500,
  k = 1,
  sequencing_depth = 1e+05,
  proportion_da = 0.1,
  size_factor_correlation = 0,
  spike_in = FALSE,
  possible_fold_changes = NULL
)
}
\arguments{
\item{p}{number of genes}

\item{n}{number of samples per group}

\item{k}{number of distinct cell types}

\item{sequencing_depth}{average total nUMI to simulate}

\item{proportion_da}{proportion of differentially abundant genes}

\item{size_factor_correlation}{correlation of observed abundances to original total abundances}

\item{spike_in}{simulate a control spike in with low dispersion}

\item{possible_fold_changes}{if specified, a distribution of possible fold changes available for differential abundance simulation}
}
\value{
named list of true abundances, observed abundances, and group assignments
}
\description{
Simulate UMI count single-cell RNA-seq data
}
\details{
Samples the mean gene abundances from Halpern et al. (2017) as a references. This gives a data set
with on average 75% zeros which may still not be representative in terms of sparsity. Need to follow up. (7/22/2020)
}
