% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spike_in_ones}
\alias{spike_in_ones}
\title{Utility function to spike-in a one-count to prevent any fully unobserved taxa}
\usage{
spike_in_ones(counts, groups = NULL)
}
\arguments{
\item{counts}{count table (samples x features)}

\item{groups}{optional group labels; if included, a one is spiked into any
feature with all-zero observations in a given group/condition}
}
\value{
counts with a minimum single observation of each feature
}
\description{
Utility function to spike-in a one-count to prevent any fully unobserved taxa
}
